Addition of ActionListners, use of intern classes and anonymous classes.
However a problem with threads and infinite loop arose. I still don't
know how to fix this, but as everything else, it shall be fixed.
I know the answer lie somewhere between Middle-Earth and the Mordor, but
my precious won't elude me for long. No, not for long...My Precious...